Improve code quality in all packages
The Dialogix project has some code quality issues, including inconsistent formatting, long methods, and missing JavaDoc comments.

These code quality issues make the project hard to maintain and can lead to bugs and readability problems.

- Refactored the code to break down long methods into smaller, more manageable functions.
- Added JavaDoc comments to enhance code documentation and readability.
- Improved formatting to adhere to coding standards.

These changes are made to enhance code maintainability, readability, and overall quality. The refactoring and documentation improvements will make the codebase more understandable and maintainable for future development.
