--- src/main/java/command/Command.java
@@ -1,18 +1,0 @@
-package command;
-
-
-import exception.DialogixException;
-import main.Storage;
-import main.TaskList;
-import main.Ui;
-
-public class Command {
-    public void execute(TaskList tasks, Ui ui, Storage storage) throws DialogixException {
-
-    }
-
-    public boolean isExit() {
-        return false;
-    }
-}
-
