--- src/main/java/command/ExitCommand.java
@@ -1,18 +1,0 @@
-package command;
-
-import main.Storage;
-import main.TaskList;
-import main.Ui;
-
-public class ExitCommand extends Command {
-
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) {
-        ui.printExitMessage();
-    }
-
-    @Override
-    public boolean isExit() {
-        return true;
-    }
-}
