+++ src/main/java/dialogix/command/Command.java
@@ -1,0 +1,31 @@
+package dialogix.command;
+
+import dialogix.exception.DialogixException;
+import dialogix.main.Storage;
+import dialogix.main.TaskList;
+import dialogix.main.Ui;
+
+/**
+ * Represents a base class for all commands in the Dialogix application.
+ */
+public abstract class Command {
+
+    /**
+     * Executes the command with the given task list, user interface, and storage.
+     *
+     * @param tasks   The task list.
+     * @param ui      The user interface.
+     * @param storage The storage manager.
+     * @throws DialogixException If there is an issue executing the command.
+     */
+    public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws DialogixException;
+
+    /**
+     * Checks if the command is an exit command.
+     *
+     * @return True if the command is an exit command, false otherwise.
+     */
+    public boolean isExit() {
+        return false;
+    }
+}
