+++ src/main/java/dialogix/command/ListCommand.java
@@ -1,0 +1,24 @@
+package dialogix.command;
+
+import dialogix.exception.DialogixException;
+import dialogix.main.Storage;
+import dialogix.main.TaskList;
+import dialogix.main.Ui;
+
+/**
+ * Represents a command to list all tasks in the task list.
+ */
+public class ListCommand extends Command {
+    /**
+     * Executes the ListCommand by displaying all tasks in the task list.
+     *
+     * @param tasks   The task list.
+     * @param ui      The user interface.
+     * @param storage The storage manager.
+     * @throws DialogixException If there is an issue executing the command.
+     */
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws DialogixException {
+        ui.printList(tasks.getAllTasks(), "Here are the tasks in your list:");
+    }
+}
