+++ src/main/java/dialogix/main/Dialogix.java
@@ -1,35 +1,22 @@
-package main;
+package dialogix.main;
 
-import command.Command;
-import exception.DialogixException;
-import task.Deadline;
-import task.Event;
-import task.Task;
-import task.Todo;
+import dialogix.command.Command;
+import dialogix.exception.DialogixException;
 
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Scanner;
-
-import javafx.application.Application;
-import javafx.scene.Scene;
-import javafx.scene.control.Label;
-import javafx.stage.Stage;
-
+/**
+ * Represents the main class for the Dialogix application.
+ */
 public class Dialogix {
-    private Storage storage;
-    private TaskList tasks;
-    private Ui ui;
-
-    private boolean isExit = false;
     private static final String filePath = "/data/dialogix.txt";
+    private final Storage storage;
+    private TaskList tasks;
+    private final Ui ui;
+    private boolean isExit = false;
 
-    Dialogix() {
+    /**
+     * Initializes the Dialogix application.
+     */
+    public Dialogix() {
         storage = new Storage(filePath);
         ui = new Ui();
 
@@ -41,7 +28,13 @@ public class Dialogix {
         }
     }
 
-    String getResponse(String input) {
+    /**
+     * Gets a response from the Dialogix application based on user input.
+     *
+     * @param input The user's input.
+     * @return The response generated by the application.
+     */
+    public String getResponse(String input) {
         if (isExit) {
             return "";
         }
