+++ src/main/java/dialogix/main/Main.java
@@ -1,4 +1,6 @@
-package main;
+package dialogix.main;
+
+import java.io.IOException;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
@@ -6,16 +8,22 @@ import javafx.scene.Scene;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
 
-import java.io.IOException;
-
+/**
+ * A GUI for the Dialogix application using FXML.
+ */
 public class Main extends Application {
 
-    protected Dialogix dialogix = new Dialogix();
+    private final Dialogix dialogix = new Dialogix();
 
+    /**
+     * Starts the JavaFX application.
+     *
+     * @param stage The primary stage for the JavaFX application.
+     */
     @Override
     public void start(Stage stage) {
         try {
-            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("MainWindow.fxml"));
+            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/view/MainWindow.fxml"));
             AnchorPane ap = fxmlLoader.load();
             Scene scene = new Scene(ap);
             stage.setScene(scene);
@@ -25,5 +33,8 @@ public class Main extends Application {
         } catch (IOException e) {
             e.printStackTrace();
         }
+    }
+    public static void main(String[] args) {
+        launch(args);
     }
 }
