--- src/main/java/exception/DialogixException.java
@@ -1,9 +1,0 @@
-package exception;
-
-public class DialogixException extends Exception {
-
-    public DialogixException(String message) {
-        super(message);
-    }
-
-}
