--- src/main/java/task/Event.java
@@ -1,34 +1,0 @@
-package task;
-
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-import java.util.Date;
-
-public class Event extends Task {
-    protected LocalDate from;
-    protected LocalDate to;
-
-    public Event(String description, LocalDate from, LocalDate to) {
-        super(description);
-        this.from = from;
-        this.to = to;
-    }
-
-    public Event(String description, Date date) {
-        super(description);
-    }
-
-    public Event(String description, String s) {
-        super(description);
-    }
-
-    @Override
-    public String toString() {
-        return "[E]" + super.toString() + " (from: " + from.format(DateTimeFormatter.ofPattern("MMM d yyyy")) + " to: " + to.format(DateTimeFormatter.ofPattern("MMM d yyyy")) + ")";
-    }
-
-    @Override
-    public String toSaveString() {
-        return "E | " + (isDone ? "1" : "0") + " | " + description + " | " + from + " | " + to;
-    }
-}
