--- src/main/java/task/TaskType.java
@@ -1,6 +1,0 @@
-package task;
-
-
-public enum TaskType {
-    TODO, DEADLINE, EVENT
-}
