--- src/main/java/task/Todo.java
@@ -1,20 +1,0 @@
-package task;
-
-public class Todo extends Task {
-
-    private boolean isDone;
-
-    public Todo(String description) {
-        super(description);
-    }
-
-    public Todo(String description, boolean isDone) {
-        super(description);
-        this.isDone = isDone;
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
