--- src/test/java/StorageTest.java
@@ -1,26 +1,0 @@
-package dialogix.main;
-
-import exception.DialogixException;
-import task.Task;
-import org.junit.jupiter.api.Test;
-
-import java.util.List;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-
-public class StorageTest {
-
-    @Test
-    void testLoadTasks() throws DialogixException {
-        // Write tests to check if tasks are loaded correctly
-        // after saving tasks to a file and then loading them
-    }
-
-    @Test
-    void testSaveTasks() throws DialogixException {
-        // Write tests to check if tasks are saved correctly to a file
-    }
-
-    // Add more tests for other methods in Storage class
-}
