+++ text-ui-test/runtest.bat
@@ -4,10 +4,10 @@ REM create bin directory if it doesn't exist
 if not exist ..\bin mkdir ..\bin
 
 REM delete output from previous run
-if exist ACTUAL.TXT del ACTUAL.TXT
+del ACTUAL.TXT
 
 REM compile the code into the bin folder
-javac  -cp ..\src\main\java\main -Xlint:none -d ..\bin ..\src\main\java\main\*.java
+javac  -cp ..\src -Xlint:none -d ..\bin ..\src\main\java\dialogix\main\*.java ..\src\main\java\dialogix\task\*.java ..\src\main\java\dialogix\exception\*.java ..\src\main\java\dialogix\command\*.java
 IF ERRORLEVEL 1 (
     echo ********** BUILD FAILURE **********
     exit /b 1
@@ -15,7 +15,8 @@ IF ERRORLEVEL 1 (
 REM no error here, errorlevel == 0
 
 REM run the program, feed commands from input.txt file and redirect the output to the ACTUAL.TXT
-java -classpath ..\bin main.Dialogix < input.txt > ACTUAL.TXT
+java -classpath ..\bin Dialogix < input.txt > ACTUAL.TXT
 
 REM compare the output to the expected output
 FC ACTUAL.TXT EXPECTED.TXT
+PAUSE
